clean temporaries
git-svn-id: https://svn.mcs.anl.gov/repos/performance/orio@1075 e7ad4b5f-b827-0410-872f-f7f4bc3d1efb

+++ orio/main/tuner/ptest_driver.py
@@ -332,7 +332,7 @@ class PerfTestDriver:
         if Globals().keep_temps: return
         for fname in [self.exe_name, self.src_name, self.timer_file]:
             try:
-                if os.path.exists(fname):
+                if fname and os.path.exists(fname):
                     os.unlink(fname)
             except:
                 err('orio.main.tuner.ptest_driver:  cannot delete file: %s' % fname)
